Gather top interfaces in single header
